rename imports of desi_common to catalog
