+++ py/legacyanalysis/euclid.py
@@ -25,7 +25,7 @@ from astrometry.util.multiproc import multiproc
 from legacypipe.runbrick import run_brick, rgbkwargs, rgbkwargs_resid
 from legacypipe.survey import LegacySurveyData, imsave_jpeg, get_rgb
 from legacypipe.image import LegacySurveyImage
-from legacypipe.desi_common import read_fits_catalog
+from legacypipe.catalog import read_fits_catalog
 
 from tractor.sky import ConstantSky
 from tractor.sfd import SFDMap
