+++ py/legacyanalysis/lsb.py
@@ -17,7 +17,7 @@ from astrometry.util.ttime import *
 from tractor import *
 
 from legacypipe.survey import *
-from legacypipe.desi_common import *
+from legacypipe.catalog import *
 
 def bin_image(data, S):
     # rebin image data
