+++ py/legacypipe/forced_photom_decam.py
@@ -19,7 +19,7 @@ from tractor.galaxy import disable_galaxy_cache
 from tractor.ellipses import EllipseE
 
 from legacypipe.survey import LegacySurveyData, bricks_touching_wcs, exposure_metadata, get_version_header, apertures_arcsec
-from desi_common import read_fits_catalog
+from catalog import read_fits_catalog
 
 # python projects/desi/forced-photom-decam.py decals/images/decam/CP20140810_g_v2/c4d_140816_032035_ooi_g_v2.fits.fz 43 DR1 f.fits
